+++ examples/counter.rs
@@ -1,71 +1,50 @@
-#![forbid(unsafe_code)]
+use async_std::task::sleep;
+use core::future::Future;
+use core::task::Poll;
+use core::time::Duration;
+use pasts::{Loop, Past};
 
-use async_std::task;
-use core::{
-    future::Future,
-    pin::Pin,
-    task::{Context, Poll},
-    time::Duration,
-};
-use pasts::{exec, wait};
+// Exit type for State.
+type Exit = ();
 
-/// An event handled by the event loop.
-enum Event {
-    One(()),
-    Two(()),
+// Shared state between tasks on the thread.
+struct State<A: Future<Output = ()>, B: Future<Output = ()>> {
+    counter: usize,
+    one: Past<(), (), A>,
+    two: Past<(), (), B>,
 }
 
-/// Shared state between tasks on the thread.
-struct State(usize);
-
-impl State {
-    /// Event loop.  Return false to stop program.
-    fn event(&mut self, event: Event) {
-        match event {
-            Event::One(()) => {
-                println!("One {}", self.0);
-                self.0 += 1;
-                if self.0 > 5 {
-                    std::process::exit(0);
-                }
-            }
-            Event::Two(()) => {
-                println!("Two {}", self.0);
-                self.0 += 1
-            }
+impl<A: Future<Output = ()>, B: Future<Output = ()>> State<A, B> {
+    fn one(&mut self, _: ()) -> Poll<Exit> {
+        println!("One {}", self.counter);
+        self.counter += 1;
+        if self.counter > 6 {
+            Poll::Ready(())
+        } else {
+            Poll::Pending
         }
     }
-}
 
-struct Interval(Duration, Pin<Box<dyn Future<Output = ()>>>);
-
-impl Interval {
-    fn new(duration: Duration) -> Self {
-        Interval(duration, Box::pin(task::sleep(duration)))
+    fn two(&mut self, _: ()) -> Poll<Exit> {
+        println!("Two {}", self.counter);
+        self.counter += 1;
+        Poll::Pending
     }
 }
 
-impl Future for &mut Interval {
-    type Output = ();
+async fn run() {
+    let mut state = State {
+        counter: 0,
+        one: Past::new((), |()| sleep(Duration::from_secs_f64(1.0))),
+        two: Past::new((), |()| sleep(Duration::from_secs_f64(2.0))),
+    };
 
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<()> {
-        match self.1.as_mut().poll(cx) {
-            Poll::Pending => Poll::Pending,
-            Poll::Ready(()) => {
-                self.1 = Box::pin(task::sleep(self.0));
-                Poll::Ready(())
-            }
-        }
-    }
+    Loop::new(&mut state)
+        .when(|s| &mut s.one, State::one)
+        .when(|s| &mut s.two, State::two)
+        .await;
 }
 
 fn main() {
-    let mut state = State(0);
-    let mut one = Interval::new(Duration::from_secs_f64(0.999));
-    let mut two = Interval::new(Duration::from_secs_f64(2.0));
-
-    exec!(state.event(wait! {
-        Event::One((&mut one).await),
-        Event::Two((&mut two).await),
-    }))
+    pasts::block_on(run())
 }