+++ examples/cross-platform/src/main.rs
@@ -1,0 +1,7 @@
+use devout::{log, Tag};
+
+const INFO: Tag = Tag::new("Info").show(true);
+
+async fn main() {
+    log!(INFO, "Hello, world!");
+}